Refactor from message polling to content message handling
