+++ js/modules/loki_primitives.d.ts
@@ -1,0 +1,11 @@
+export async function sleepFor(ms: number);
+export async function allowOnlyOneAtATime(
+  name: any,
+  process: any,
+  timeout?: any
+);
+
+export async function abortableIterator(
+  array: Array<any>,
+  action: (any) => void
+);
