+++ js/modules/metadata/SecretSessionCipher.js
@@ -326,6 +326,7 @@ SecretSessionCipher.prototype = {
     const _decryptWithSecretKeys = this._decryptWithSecretKeys.bind(this);
 
     const ourIdentity = await signalProtocolStore.getIdentityKeyPair();
+
     const wrapper = _createUnidentifiedSenderMessageFromBuffer(ciphertext);
     const ephemeralSalt = concatenateBytes(
       bytesFromString(UNIDENTIFIED_DELIVERY_PREFIX),
