+++ package.json
@@ -35,6 +35,7 @@
     "test-node": "mocha --recursive --exit test/app test/modules ts/test libloki/test/node --timeout 10000",
     "test-session": "mocha --recursive --exit ts/test/session --timeout 10000",
     "test-medium-groups": "ELECTRON_DISABLE_SANDBOX=1 mocha --exit --timeout 10000 integration_test/integration_test.js --grep 'senderkeys'",
+    "test-groups": "ELECTRON_DISABLE_SANDBOX=1 mocha --exit --timeout 10000 integration_test/integration_test.js --grep 'Closed groups'",
     "eslint": "eslint --cache .",
     "eslint-fix": "eslint --fix .",
     "eslint-full": "eslint .",
@@ -133,6 +134,7 @@
     "@types/lodash": "4.14.106",
     "@types/mkdirp": "0.5.2",
     "@types/mocha": "5.0.0",
+    "@types/node-fetch": "^2.5.7",
     "@types/pify": "3.0.2",
     "@types/qs": "6.5.1",
     "@types/rc-slider": "^8.6.5",
