+++ ts/receiver/attachments.ts
@@ -1,5 +1,7 @@
 import { MessageModel } from '../../js/models/messages';
 
+import * as Data from '../../js/modules/data';
+
 // TODO: Might convert it to a class later
 let webAPI: any;
 
@@ -230,7 +232,7 @@ export async function queueAttachmentDownloads(
   }
 
   if (count > 0) {
-    await window.Signal.Data.saveMessage(message.attributes, {
+    await Data.saveMessage(message.attributes, {
       Message: Whisper.Message,
     });
